Target .NET Core 2.1 and .NET Framework 3.5
