+++ De4DotCommon.props
@@ -1,0 +1,26 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+
+  <PropertyGroup>
+    <De4DotNetFramework>false</De4DotNetFramework>
+    <De4DotNetFramework Condition=" '$(SolutionFileName)' == 'de4dot.netframework.sln' ">true</De4DotNetFramework>
+    <!-- Two different sln files are used because some of the projects are only available when targetting .NET Framework -->
+    <TargetFrameworks Condition=" '$(De4DotNetFramework)' == 'true' ">net35</TargetFrameworks>
+    <TargetFrameworks Condition=" '$(De4DotNetFramework)' != 'true' ">netcoreapp2.1</TargetFrameworks>
+    <Features>strict</Features>
+    <LangVersion>latest</LangVersion>
+    <SignAssembly>true</SignAssembly>
+    <AssemblyOriginatorKeyFile>$(MSBuildThisFileDirectory)\de4dot.snk</AssemblyOriginatorKeyFile>
+    <Version>3.1.41592.3405</Version>
+    <Copyright>Copyright (C) 2011-2018 de4dot@gmail.com</Copyright>
+    <OutputPath>$(MSBuildThisFileDirectory)\$(Configuration)</OutputPath>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
+    <DefineConstants>$(DefineConstants);DEBUG;TRACE</DefineConstants>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
+    <DefineConstants>$(DefineConstants);TRACE</DefineConstants>
+  </PropertyGroup>
+
+</Project>
