+++ Test.Rename.Dll/Tests.cs
@@ -21,6 +21,8 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 
+#pragma warning disable CS0693, CS0414, CS0169, CS0108, CS0067
+
 namespace Test.Rename.Dll {
 	static class g {
 		static public T m<T>(T t) {
