+++ de4dot.code/AssemblyClient/AssemblyClientFactory.cs
@@ -30,7 +30,12 @@ namespace de4dot.code.AssemblyClient {
 	}
 
 	public class NewAppDomainAssemblyClientFactory : IAssemblyClientFactory {
-		public IAssemblyClient Create(AssemblyServiceType serviceType) => new AssemblyClient(new NewAppDomainAssemblyServerLoader(serviceType));
+		public IAssemblyClient Create(AssemblyServiceType serviceType) =>
+#if NET35
+			new AssemblyClient(new NewAppDomainAssemblyServerLoader(serviceType));
+#else
+			new AssemblyClient(new SameAppDomainAssemblyServerLoader(serviceType));
+#endif
 	}
 
 	public class NewProcessAssemblyClientFactory : IAssemblyClientFactory {
@@ -40,10 +45,18 @@ namespace de4dot.code.AssemblyClient {
 		public NewProcessAssemblyClientFactory(ServerClrVersion serverVersion) => this.serverVersion = serverVersion;
 
 		public IAssemblyClient Create(AssemblyServiceType serviceType, ModuleDef module) =>
+#if NET35
 			new AssemblyClient(new NewProcessAssemblyServerLoader(serviceType, GetServerClrVersion(module)));
+#else
+			new AssemblyClient(new SameAppDomainAssemblyServerLoader(serviceType));
+#endif
 
 		public IAssemblyClient Create(AssemblyServiceType serviceType) =>
+#if NET35
 			new AssemblyClient(new NewProcessAssemblyServerLoader(serviceType, serverVersion));
+#else
+			new AssemblyClient(new SameAppDomainAssemblyServerLoader(serviceType));
+#endif
 
 		public static ServerClrVersion GetServerClrVersion(ModuleDef module) {
 			switch (module.GetPointerSize()) {
