+++ de4dot.code/AssemblyClient/NewAppDomainAssemblyServerLoader.cs
@@ -17,6 +17,7 @@
     along with de4dot.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+#if NET35
 using System;
 using System.Threading;
 using AssemblyData;
@@ -84,3 +85,4 @@ namespace de4dot.code.AssemblyClient {
 		}
 	}
 }
+#endif
