+++ de4dot.code/AssemblyClient/ServerClrVersion.cs
@@ -17,17 +17,14 @@
     along with de4dot.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("de4dot.blocks")]
-[assembly: AssemblyDescription("Modifies dnlib MethodDef bodies")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("de4dot.blocks")]
-[assembly: AssemblyCopyright("Copyright (C) 2011-2015 de4dot@gmail.com")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: ComVisible(false)]
-[assembly: AssemblyVersion("3.1.41592.3405")]
-[assembly: AssemblyFileVersion("3.1.41592.3405")]
+namespace de4dot.code.AssemblyClient {
+	public enum ServerClrVersion {
+		CLR_ANY_ANYCPU,
+		CLR_ANY_x86,
+		CLR_ANY_x64,
+		CLR_v20_x86,
+		CLR_v20_x64,
+		CLR_v40_x86,
+		CLR_v40_x64,
+	}
+}
