+++ de4dot.cui/Properties/AssemblyInfo.cs
@@ -17,20 +17,7 @@
     along with de4dot.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("de4dot.cui")]
-[assembly: AssemblyDescription("de4dot Console UI code")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("de4dot.cui")]
-[assembly: AssemblyCopyright("Copyright (C) 2011-2015 de4dot@gmail.com")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: ComVisible(false)]
-[assembly: AssemblyVersion("3.1.41592.3405")]
-[assembly: AssemblyFileVersion("3.1.41592.3405")]
 [assembly: InternalsVisibleTo("de4dot, PublicKey=00240000048000009400000006020000002400005253413100040000010001006dc721ade368b747b89b536e7f1887993469726dd10be2ed8527a5b9c473633afe734730bd801e2e888f4462ce9b64a17df849453af430351ade9b4e5b1071fcd7f036871e85c6be368fe65b750b87e44dbea709b0524de642803c31db2a6ac68a61bb09e9289b14ce372190a50b94d7bc021366d401d53d54f6e1907259cda9")]
 [assembly: InternalsVisibleTo("de4dot-x64, PublicKey=00240000048000009400000006020000002400005253413100040000010001006dc721ade368b747b89b536e7f1887993469726dd10be2ed8527a5b9c473633afe734730bd801e2e888f4462ce9b64a17df849453af430351ade9b4e5b1071fcd7f036871e85c6be368fe65b750b87e44dbea709b0524de642803c31db2a6ac68a61bb09e9289b14ce372190a50b94d7bc021366d401d53d54f6e1907259cda9")]
