+++ deobfuscator.Template/deobfuscator.Template.csproj
@@ -1,68 +1,20 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{88AC39C0-4CCF-44C4-9C24-301459E6C0D1}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>deobfuscator.Template</RootNamespace>
-    <AssemblyName>deobfuscator.Template</AssemblyName>
-    <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
-    <TargetFrameworkProfile>Client</TargetFrameworkProfile>
-    <FileAlignment>512</FileAlignment>
+    <TargetFrameworks>net35;netcoreapp2.1</TargetFrameworks>
     <SignAssembly>true</SignAssembly>
     <AssemblyOriginatorKeyFile>..\de4dot.snk</AssemblyOriginatorKeyFile>
     <Features>strict</Features>
     <LangVersion>latest</LangVersion>
+    <OutputPath>..\$(Configuration)</OutputPath>
   </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>portable</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>..\Debug\bin\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>portable</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>..\Release\bin\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
+
   <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Xml" />
+    <ProjectReference Include="..\AssemblyData\AssemblyData.csproj" />
+    <ProjectReference Include="..\de4dot.blocks\de4dot.blocks.csproj" />
+    <ProjectReference Include="..\de4dot.code\de4dot.code.csproj" />
+    <ProjectReference Include="..\de4dot.mdecrypt\de4dot.mdecrypt.csproj" />
+    <ProjectReference Include="..\dnlib\src\dnlib.csproj" />
   </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\AssemblyData\AssemblyData.csproj">
-      <Project>{FBD84077-9D35-41FE-89DF-8D79EFE0B595}</Project>
-      <Name>AssemblyData</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\de4dot.blocks\de4dot.blocks.csproj">
-      <Project>{045B96F2-AF80-4C4C-8D27-E38635AC705E}</Project>
-      <Name>de4dot.blocks</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\de4dot.code\de4dot.code.csproj">
-      <Project>{4D10B9EB-3BF1-4D61-A389-CB019E8C9622}</Project>
-      <Name>de4dot.code</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\de4dot.mdecrypt\de4dot.mdecrypt.csproj">
-      <Project>{5C93C5E2-196F-4877-BF65-96FEBFCEFCA1}</Project>
-      <Name>de4dot.mdecrypt</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\dnlib\src\dnlib.csproj">
-      <Project>{FDFC1237-143F-4919-8318-4926901F4639}</Project>
-      <Name>dnlib</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+
 </Project>
